+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dockerfile
 - added JSDoc for functions
 - added warnings about pre-production usage and simple roadmap about versioning and release schedule
+- concat function
 
 ### Changed
 - added more types